requirements.txt
Flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
python-socketio==5.10.0
eventlet==0.33.3
psutil==5.9.6
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
